changes requested by cppwfs
