+++ spring-cloud-task-stream/src/main/java/org/springframework/cloud/task/batch/listener/EventEmittingItemReadListener.java
@@ -49,7 +49,7 @@ public class EventEmittingItemReadListener implements ItemReadListener, Ordered 
 		this.messagePublisher = new MessagePublisher(output);
 	}
 
-	public EventEmittingItemReadListener(MessageChannel output,int order) {
+	public EventEmittingItemReadListener(MessageChannel output, int order) {
 		Assert.notNull(output, "An output channel is required");
 		this.messagePublisher = new MessagePublisher(output);
 		this.order = order;
