+++ spring-cloud-task-stream/src/main/java/org/springframework/cloud/task/batch/listener/EventEmittingItemWriteListener.java
@@ -50,7 +50,7 @@ public class EventEmittingItemWriteListener implements ItemWriteListener, Ordere
 		this.messagePublisher = new MessagePublisher<>(output);
 	}
 
-	public EventEmittingItemWriteListener(MessageChannel output,int order) {
+	public EventEmittingItemWriteListener(MessageChannel output, int order) {
 		Assert.notNull(output, "An output channel is required");
 		this.messagePublisher = new MessagePublisher<>(output);
 		this.order = order;
