+++ spring-cloud-task-stream/src/main/java/org/springframework/cloud/task/batch/listener/EventEmittingJobExecutionListener.java
@@ -40,7 +40,7 @@ public class EventEmittingJobExecutionListener implements JobExecutionListener, 
 		this.messagePublisher = new MessagePublisher<>(output);
 	}
 
-	public EventEmittingJobExecutionListener(MessageChannel output,int order) {
+	public EventEmittingJobExecutionListener(MessageChannel output, int order) {
 		Assert.notNull(output, "An output channel is required");
 		this.messagePublisher = new MessagePublisher<>(output);
 		this.order = order;
