+++ spring-cloud-task-stream/src/main/java/org/springframework/cloud/task/batch/listener/support/TaskEventProperties.java
@@ -1,3 +1,19 @@
+/*
+ * Copyright 2017 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package org.springframework.cloud.task.batch.listener.support;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
@@ -6,88 +22,98 @@ import org.springframework.core.Ordered;
 /**
  *  @author Ali Shahbour
  */
-@ConfigurationProperties(prefix = "spring.cloud.task.events")
+@ConfigurationProperties(prefix = "spring.cloud.task.batch.events")
 public class TaskEventProperties {
 
-    private ListenerProperties jobExecution = new ListenerProperties(Ordered.HIGHEST_PRECEDENCE);
-    private ListenerProperties stepExecution = new ListenerProperties(Ordered.HIGHEST_PRECEDENCE);
-    private ListenerProperties itemRead = new ListenerProperties(Ordered.HIGHEST_PRECEDENCE);
-    private ListenerProperties itemProcess = new ListenerProperties(Ordered.HIGHEST_PRECEDENCE);
-    private ListenerProperties itemWrite = new ListenerProperties(Ordered.HIGHEST_PRECEDENCE);
-    private ListenerProperties chunk = new ListenerProperties(Ordered.HIGHEST_PRECEDENCE);
-    private ListenerProperties skip = new ListenerProperties(Ordered.HIGHEST_PRECEDENCE);
+    /**
+     * Properties for jobExecution listener
+     */
+	private ListenerProperties jobExecution = new ListenerProperties(Ordered.HIGHEST_PRECEDENCE);
 
-    public ListenerProperties getJobExecution() {
-        return jobExecution;
-    }
+    /**
+     * Properties for stepExecution listener
+     */
+	private ListenerProperties stepExecution = new ListenerProperties(Ordered.HIGHEST_PRECEDENCE);
 
-    public void setJobExecution(ListenerProperties jobExecution) {
-        this.jobExecution = jobExecution;
-    }
+    /**
+     * Properties for itemRead listener
+     */
+	private ListenerProperties itemRead = new ListenerProperties(Ordered.HIGHEST_PRECEDENCE);
 
-    public ListenerProperties getStepExecution() {
-        return stepExecution;
-    }
+    /**
+     * Properties for itemProcess listener
+     */
+	private ListenerProperties itemProcess = new ListenerProperties(Ordered.HIGHEST_PRECEDENCE);
 
-    public void setStepExecution(ListenerProperties stepExecution) {
-        this.stepExecution = stepExecution;
-    }
+    /**
+     * Properties for itemWrite listener
+     */
+	private ListenerProperties itemWrite = new ListenerProperties(Ordered.HIGHEST_PRECEDENCE);
 
-    public ListenerProperties getItemRead() {
-        return itemRead;
-    }
+    /**
+     * Properties for chunk listener
+     */
+	private ListenerProperties chunk = new ListenerProperties(Ordered.HIGHEST_PRECEDENCE);
 
-    public void setItemRead(ListenerProperties itemRead) {
-        this.itemRead = itemRead;
-    }
+    /**
+     * Properties for skip listener
+     */
+	private ListenerProperties skip = new ListenerProperties(Ordered.HIGHEST_PRECEDENCE);
 
-    public ListenerProperties getItemProcess() {
-        return itemProcess;
-    }
+	public ListenerProperties getJobExecution() {
+		return jobExecution;
+	}
 
-    public void setItemProcess(ListenerProperties itemProcess) {
-        this.itemProcess = itemProcess;
-    }
+	public void setJobExecution(ListenerProperties jobExecution) {
+		this.jobExecution = jobExecution;
+	}
 
-    public ListenerProperties getItemWrite() {
-        return itemWrite;
-    }
+	public ListenerProperties getStepExecution() {
+		return stepExecution;
+	}
 
-    public void setItemWrite(ListenerProperties itemWrite) {
-        this.itemWrite = itemWrite;
-    }
+	public void setStepExecution(ListenerProperties stepExecution) {
+		this.stepExecution = stepExecution;
+	}
 
-    public ListenerProperties getChunk() {
-        return chunk;
-    }
+	public ListenerProperties getItemRead() {
+		return itemRead;
+	}
 
-    public void setChunk(ListenerProperties chunk) {
-        this.chunk = chunk;
-    }
+	public void setItemRead(ListenerProperties itemRead) {
+		this.itemRead = itemRead;
+	}
 
-    public ListenerProperties getSkip() {
-        return skip;
-    }
+	public ListenerProperties getItemProcess() {
+		return itemProcess;
+	}
 
-    public void setSkip(ListenerProperties skip) {
-        this.skip = skip;
-    }
+	public void setItemProcess(ListenerProperties itemProcess) {
+		this.itemProcess = itemProcess;
+	}
 
-    public  class ListenerProperties {
+	public ListenerProperties getItemWrite() {
+		return itemWrite;
+	}
 
+	public void setItemWrite(ListenerProperties itemWrite) {
+		this.itemWrite = itemWrite;
+	}
 
-        public ListenerProperties(int order) {
-            this.order = order;
-        }
+	public ListenerProperties getChunk() {
+		return chunk;
+	}
 
-        private int order;
+	public void setChunk(ListenerProperties chunk) {
+		this.chunk = chunk;
+	}
 
-        public int getOrder() {
-            return order;
-        }
+	public ListenerProperties getSkip() {
+		return skip;
+	}
 
-        public void setOrder(int order) {
-            this.order = order;
-        }
-    }
+	public void setSkip(ListenerProperties skip) {
+		this.skip = skip;
+	}
+
 }
